# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DiminishStudio API - Music learning and chord analysis platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SongDifficulty = typeof SongDifficulty[keyof typeof SongDifficulty];


export const SongDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Song {
  id: number;
  title: string;
  artist: string;
  genre: string;
  /** Duration in seconds */
  duration: number;
  /** @nullable */
  coverUrl: string | null;
  bpm: number;
  key: string;
  difficulty: SongDifficulty;
  playCount?: number;
}

export interface LyricLine {
  /** Time in seconds */
  time: number;
  text: string;
  chords: string[];
}

export interface ChordMeasure {
  measure: number;
  beat: number;
  chord: string;
  time: number;
}

export type AudioTrackInstrument = typeof AudioTrackInstrument[keyof typeof AudioTrackInstrument];


export const AudioTrackInstrument = {
  guitar: 'guitar',
  piano: 'piano',
  bass: 'bass',
  drums: 'drums',
  vocal: 'vocal',
  synth: 'synth',
} as const;

export interface AudioTrack {
  id: number;
  instrument: AudioTrackInstrument;
  label: string;
  /** 0-100 */
  volume: number;
  muted: boolean;
}

export interface SongDetail {
  id: number;
  title: string;
  artist: string;
  genre: string;
  duration: number;
  /** @nullable */
  coverUrl: string | null;
  bpm: number;
  key: string;
  difficulty: string;
  lyrics: LyricLine[];
  chordTimeline: ChordMeasure[];
  tracks: AudioTrack[];
}

export interface ProcessSongInput {
  /** URL or "upload" */
  source: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  artist?: string | null;
}

export type ProcessingJobStatus = typeof ProcessingJobStatus[keyof typeof ProcessingJobStatus];


export const ProcessingJobStatus = {
  queued: 'queued',
  analyzing: 'analyzing',
  extracting_chords: 'extracting_chords',
  syncing_lyrics: 'syncing_lyrics',
  finalizing: 'finalizing',
  done: 'done',
  error: 'error',
} as const;

export interface ProcessingJob {
  jobId: string;
  status: ProcessingJobStatus;
  /** 0-100 */
  progress: number;
  /** @nullable */
  songId?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export interface Chord {
  id: number;
  name: string;
  instrument: string;
  difficulty: string;
  fingers: number[];
  strings: number[];
  /** @nullable */
  description?: string | null;
  /** @nullable */
  audioUrl?: string | null;
}

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  preferredInstrument?: string | null;
  createdAt: string;
}

export interface UserUpdate {
  username?: string;
  bio?: string;
  preferredInstrument?: string;
}

export interface RegisterInput {
  username: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface LibraryEntry {
  id: number;
  songId: number;
  addedAt: string;
  song: Song;
}

export interface StorageQuota {
  /** Total bytes used by user's uploaded files */
  storageUsedBytes: number;
  /** Maximum bytes allowed for the user (configurable per-user) */
  storageQuotaBytes: number;
}

export interface LibraryInput {
  songId: number;
}

export interface LearningSession {
  id: number;
  chordName: string;
  instrument: string;
  startedAt: string;
  attemptsCount: number;
  successCount: number;
  mastered?: boolean;
}

export interface LearningSessionInput {
  chordName: string;
  instrument: string;
}

export interface ChordAttemptInput {
  success: boolean;
  /** 0-1 */
  confidenceScore: number;
}

export interface ChordAttempt {
  id: number;
  sessionId: number;
  success: boolean;
  confidenceScore: number;
  attemptedAt: string;
}

export interface MasteredChord {
  chordName: string;
  instrument: string;
  masteredAt: string;
  totalAttempts: number;
  successRate: number;
}

export type ListSongsParams = {
genre?: string;
search?: string;
};

export type ListChordsParams = {
instrument?: ListChordsInstrument;
difficulty?: ListChordsDifficulty;
};

export type ListChordsInstrument = typeof ListChordsInstrument[keyof typeof ListChordsInstrument];


export const ListChordsInstrument = {
  guitar: 'guitar',
  piano: 'piano',
  bass: 'bass',
  ukulele: 'ukulele',
} as const;

export type ListChordsDifficulty = typeof ListChordsDifficulty[keyof typeof ListChordsDifficulty];


export const ListChordsDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;