# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DiminishStudio API - Music learning and chord analysis platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library:
    get:
      operationId: getUserLibrary
      tags: [library]
      summary: Get user's saved song library
      responses:
        "200":
          description: User library
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryEntry"
    post:
      operationId: addToLibrary
      tags: [library]
      summary: Add a song to user's library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryInput"
      responses:
        "201":
          description: Added to library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"
  /library/{songId}:
    delete:
      operationId: removeFromLibrary
      tags: [library]
      summary: Remove a song from library
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Song:
      type: object
      required: [id, title, artist, genre, duration, coverUrl, bpm, key, difficulty]
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        genre:
          type: string
        duration:
          type: integer
          description: Duration in seconds
        coverUrl:
          type: ["string", "null"]
        bpm:
          type: integer
        key:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        playCount:
          type: integer
...
    LibraryEntry:
      type: object
      required: [id, songId, addedAt, song]
      properties:
        id:
          type: integer
        songId:
          type: integer
        addedAt:
          type: string
        song:
          $ref: "#/components/schemas/Song"
...
    LibraryInput:
      type: object
      required: [songId]
      properties:
        songId:
          type: integer
...